Implement Template design Pattern step 1
Notice that all the code in the four Expression types we already wrote
was very similar. A perfect case for Template. So we implement a class
that describes the algorithm while deferring the choice of operation
to the derived classes. So far all we've done is write this class,
called BinaryExpression. Note that it implements Expression's
evaluate, but defines operateOn as pure virtual, so derived classes
will need to implement that.